proyecto_DeliveryPuyo/requirements.txt
Flask
Flask-Login
Flask-SQLAlchemy
mysql-connector-python
fpdf2
Werkzeug